+++ src/main/scala/io/viash/helpers/ReplayableMultiOutputStream.scala
@@ -1,3 +1,20 @@
+/*
+ * Copyright (C) 2020  Data Intuitive
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package io.viash.helpers
 
 import java.io.ByteArrayOutputStream
@@ -44,5 +61,9 @@ class ReplayableMultiOutputStream {
     buffer.foreach { case (index, data) =>
       callbacks(index)(data)
     }
+  }
+
+  override def toString(): String = {
+    buffer.map(_._2).mkString
   }
 }